refactor(mcp): rename upstream_credentials.py to resolver.py
Avoids the outbound_credentials/upstream_credentials.py redundancy; the file holds the one
resolve() entry point, so resolver.py names it for what it is. Renames the mapped test to
test_resolver.py to match, and updates the test import and the types.py docstring pointer.
Pure rename; UpstreamCredentialProvider keeps its name.
