+++ litellm/proxy/gateway/mcp/outbound_credentials/resolver.py
